# docs/swagger.json
{
    "schemes": [
        "http",
        "https"
    ],
    "swagger": "2.0",
    "info": {
        "description": "High-Performance Anime Streaming API built with Go\nA blazing-fast, production-ready RESTful API for anime streaming platforms\n\n**Features:**\n- Complete anime data (episodes, metadata, streaming sources)\n- Advanced search and filtering\n- Schedule management and notifications\n- Character and voice actor information\n- User watchlist support\n- Optimized caching with Redis\n\n**Disclaimer:**\nThis API does not store any files. It only links to media hosted on 3rd party services.\nThis API is explicitly made for educational purposes only.",
        "title": "Jutsu API",
        "contact": {
            "name": "Jutsu API Support",
            "url": "https://github.com/mnuddindev/jutsu-api",
            "email": "support@jutsu-api.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0.0"
    },
    "host": "localhost:8080",
    "basePath": "/api",
    "paths": {
...
        "/actors/{id}": {
            "get": {
                "description": "Returns detailed information about a specific voice actor and their roles",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Characters"
                ],
                "summary": "Get voice actor details",
                "parameters": [
                    {
                        "type": "string",
                        "example": "gakuto-kajiwara-534",
                        "description": "Voice actor ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Voice actor details",
                        "schema": {
                            "$ref": "#/definitions/handler.SuccessResponse"
                        }
                    },
                    "400": {
                        "description": "Missing path parameter",
                        "schema": {
                            "$ref": "#/definitions/handler.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "No voice actor found",
                        "schema": {
                            "$ref": "#/definitions/handler.ErrorResponse"
                        }
                    },
                    "502": {
                        "description": "Failed to fetch voice actor",
                        "schema": {
                            "$ref": "#/definitions/handler.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "handler.ErrorResponse": {
            "description": "Standard error response",
            "type": "object",
            "properties": {
                "message": {
                    "type": "string",
                    "example": "Error message here"
                },
                "success": {
                    "type": "boolean",
                    "example": false
                }
            }
        },
...
        "handler.SuccessResponse": {
            "description": "Standard success response",
            "type": "object",
            "properties": {
                "cached": {
                    "type": "boolean",
                    "example": false
                },
                "results": {},
                "success": {
                    "type": "boolean",
                    "example": true
                }
            }
        }
    },
...
}